delete editorial
