+++ lib/util/constants.dart
@@ -467,17 +467,12 @@ class MixpanelEvent {
   static const connectExternal = 'connect_external';
   static const connectAutonomyDisplay = 'connect_autonomy_display';
   static const subcription = 'Subcription';
-  static const editorialViewArticle = 'editorial_view_article';
-  static const editorialReadingArticle = 'editorial_reading_article';
   static const addNFT = 'add_NFT';
   static const enableNotification = 'enable_notification';
   static const tabNotification = 'tab_notification';
-  static const viewEditorial = 'view_editorial';
-  static const timeViewEditorial = 'time_view_editorial';
   static const finishArticles = 'finish_articles';
   static const visitExhibition = 'visit_exhibition';
   static const visitExhibitionArtwork = 'visit_exhibition_artwork';
-  static const tabOnLinkInEditorial = 'tab_on_link_in_editorial';
   static const createPlaylist = 'create_playlist';
   static const undoCreatePlaylist = 'undo_create_playlist';
   static const scanQR = 'scan_qr';
@@ -497,7 +492,6 @@ class MixpanelEvent {
   static const hideAddresses = "hide_addresses";
   static const callIrlFunction = "call_irl_function";
   static const numberNft = "number_nft";
-  static const editorialReadingTimeByWeek = "editorial_reading_time_by_week";
 }
 
 class MixpanelProp {
